HAI-1526 Separate and refactor usercontact
